+++ packages/eslint/src/preset-rules/typescript.ts
@@ -1,0 +1,10 @@
+import type { Linter } from 'eslint'
+
+/**
+ * - rule: `ts/*`
+ * - original rule: `@typescript-eslint/*`
+ * - plugin: `@typescript-eslint/eslint-plugin`
+ * @see https://github.com/antfu/eslint-config#typescript
+ * @see https://typescript-eslint.io/rules/
+ */
+export const TYPESCRIPT_PRESET_RULES: Linter.RulesRecord = {}